+++ yash-semantics/src/lib.rs
@@ -28,13 +28,14 @@
 
 pub mod assign;
 pub mod command;
-pub mod command_search;
 pub mod expansion;
 pub mod job;
 pub mod redir;
 pub mod trap;
 pub mod xtrace;
 
+#[doc(no_inline)]
+pub use yash_env::semantics::command::search as command_search;
 #[doc(no_inline)]
 pub use yash_env::semantics::*;